Breaking changes! See addional commit notes
  * Order of args on dyadic Inline::J::Verbs has been swapped

Under previous version, an `Inline::J::Verbs` would create 2 multi-subs

eg.
    multi sub foo($y)     { ... }
    multi sub foo($x, $y) { ... }

Since in J, all verbs are infix - and Raku can create infix ops - I
thought this would be good for simple creation of infix open, eg.

    my &infix:<op> := j.verb('SomeDyadidVerb');
    say 'x' op 'y';

While this is nice, I don't use it much, and since the "main" argument
of a J verb is typically on the RHS "y" operand, it makes sense to make
it the first argument in a Raku function.

For now, making infix ops in Raku will require explicit swapping of the
arguments, eg.

    sub infix:<op>($x, $y) { j.verb('SomeDyadicVerb')($y, $x) }

    # Or just use `~` (reflex, aka: commute)
    my &infix:<op> := j.verb('SomeDyadicVerb~');

I think the consistency in the first argument always being `y` is worth
this breaking change.

  * The `getm` method now has a `:list` option

Calling `$some-noun.getm` returns a shaped array. Unfortunately, it's
still a little difficult to work with shaped arrays in Raku, and most
users use nested lists. Calling `$some-noun.getm(:list)` on a noun that
would normally return an `Array` will now return a `List`. If has
multiple dimensions, it will return a nested list. Of course, this list
loses it's "type".

Since lists are immutable, I am thinking this should perhaps be a nested
Array instead. Also, nouns where `.getm` would  return a `Scalar` will
still do so. This begs the question, is `:list` a good name for this
option? Probably not.

  * Update "depends" in META

Just adding the `auth` identifier to be safe. That is all.