Additional IJV's changes, README updates, and more
Under a previous commit, the order of dyadic IJV's changed so that the y
argument came first, however I missed the multi's that accept IJN's.
This has now been corrected.

While still currently undocumented, the `:list` option for  `getm` and
`gets` is no more. As those methods could return scalar values, calling
it "list" was not always applicable. There is now a `:shaped` boolean
that defaults to true. To get a nested-list from `getm` that would
return a shaped array is now `.getm(:!shaped)`. Scalars are "unshaped"
so this term is more broadly applicable (if a little verbose.)

Updated the README