Always load-profile, and add more IJV candidates
I practically always call `new` with `:load-profile`, so I'm making it
the default. Use `:!load-profile` when it's not needed.

The CALL-ME candidates on IJV's now accepts Str, which are converted to
an Array, and passed to the candidate that handles Array. The Array is
converted to an IJN and passed to the candidate that handles those.