Add scalar handling in setm, simplify IJV CALL-ME
Previously, almost all Raku values passed to setm (which includes any
Raku value passed as an argument to an Inline::J::Verb) would be
converted to an array.

While J is usually ok with handling a single-element list when it
expected a single scarlar, this permeated to return values, so now
single values are not coerced to Array.

Int's a are J integers, other Real's are converced to Num's and become
J's floating type.

The other benefit to all this is less cases to handle for the CALL-ME
function of Inline::J::Verbs.